go back to dumb mock storage
