+++ src/families/bitcoininjs/blockchain/__tests__/wallet.synced.unit.test.ts
@@ -1,4 +1,4 @@
-import Storage from "../storage/orderedmemorylist";
+import Storage from "../storage/mock";
 import Explorer from "../explorer/mock";
 import Derivation from "../derivation/bitcoin";
 import Wallet from "../wallet";
